to_dict() is tricky
